Change timestamp types back to uint32 for new format

+++ zipper/protocols/v2/protobuf_group.go
@@ -144,9 +144,9 @@ func (c *ClientProtoV2Group) Fetch(ctx context.Context, request *protov3.MultiFe
 		r.Metrics = append(r.Metrics, protov3.FetchResponse{
 			Name:              m.Name,
 			ConsolidationFunc: "average",
-			StopTime:          m.StopTime,
-			StartTime:         m.StartTime,
-			StepTime:          m.StepTime,
+			StopTime:          uint32(m.StopTime),
+			StartTime:         uint32(m.StartTime),
+			StepTime:          uint32(m.StepTime),
 			Values:            m.Values,
 			XFilesFactor:      0.0,
 		})
